app/requirements.txt
Flask==2.2.5
celery==5.3.1
redis==4.5.1
pdf2image==1.16.0
pytesseract==0.3.10
python-docx==0.8.11
Pillow==10.0.0
#arabic-reshaper==2.1.0
python-bidi==0.4.1
gunicorn==20.1.0
python-dotenv==1.0.0
